int64 indexing: fix reduction idx/index variables to long
ArgMax/ArgMin/CumAdd/CumMul used int for idx/maxAt/minAt variables.
These would overflow for arrays > 2B elements.

Fixed:
- Default.Reduction.ArgMax.cs: int idx = 1, maxAt = 0 -> long
- Default.Reduction.ArgMin.cs: int idx = 1, minAt = 0 -> long
- Default.Reduction.CumAdd.cs: int idx = 0 -> long
- Default.Reduction.CumMul.cs: int idx = 0 -> long

All 3913 tests pass.

+++ src/NumSharp.Core/Backends/Default/Math/Reduction/Default.Reduction.CumAdd.cs
@@ -107,7 +107,7 @@ namespace NumSharp.Backends
 
                 // Write to output with proper type handling
                 double sum = 0;
-                int idx = 0;
+                long idx = 0;
                 while (hasNext())
                 {
                     sum += moveNext();
